Move ListItemAction -> ugly_list.rs
